separate storage and database (JsonDB)
backport of:
 - https://github.com/spesmilo/electrum/commit/7f2083f667caea442eae3184dadba4e29be7c447
 - partial https://github.com/spesmilo/electrum/commit/7f2083f667caea442eae3184dadba4e29be7c447
